requirements.txt
Flask==3.0.0
flask-sock==0.7.0
websockets==12.0
simple-websocket==1.0.0
python-dotenv==1.0.0

